Get basic networking done
efficiently send entity keyvalues needed for rendering in bspguy. I don't know how HL really does it, but I think this does roughly the same thing. Changes to entities are sent, relative to a baseline which is constantly changing as the client acks packets it receieved. The main difference is the baselines are set per update instead of per entity, which should result in less network usage. Also only the keys needed for rendering are sent.